# Slight improve on numeric parser perf