Befor:
```
BenchmarkNumberType/no-num
BenchmarkNumberType/no-num-10         	  332910	      3478 ns/op	  69.87 MB/s	    3162 B/op	      45 allocs/op
BenchmarkNumberType/num
BenchmarkNumberType/num-10            	  313408	      3839 ns/op	  63.29 MB/s	    3555 B/op	      64 allocs/op
```

After
```
BenchmarkNumberType/no-num
BenchmarkNumberType/no-num-10         	  345669	      3483 ns/op	  69.78 MB/s	    3162 B/op	      45 allocs/op
BenchmarkNumberType/num
BenchmarkNumberType/num-10            	  328444	      3728 ns/op	  65.18 MB/s	    3426 B/op	      57 allocs/op
```